Triggered Build: 24.01.003                                             Original message: Update default time in daily scheduled job
Changed the start_scheduled_job method in the scheduler_service.py file. Instead of having a fixed default time of "04:00" for daily jobs, the default time will now be dynamically set to the value of the period argument. This change allows for more flexible scheduling of daily tasks. Additionally, a conditional check has been corrected to compare the time_unit with `TimeUnit.DAILY` enum value instead of the string 'daily'.