Triggered Build: 24.02.002                                             Original message: Add library name environment variable and dependent services update stage to Jenkinsfile
This commit adds a new environment variable to specify the library name in the Jenkinsfile. It also introduces a new pipeline stage to update dependent services, which includes a waiting period and an HTTP request function to an updater URL. These changes facilitate a more dynamic interaction with the internal services.